fixing compiler errors in vs2003
git-svn-id: http://crashrpt.googlecode.com/svn@1093 821a4644-a21a-0410-aec7-6199d47568cc

+++ branches/1.2/reporting/crashsender/FilePreviewCtrl.cpp
@@ -191,7 +191,7 @@ BOOL CImage::IsJPEG(FILE* f)
 	// Each JPEG file begins with SOI marker (0xD8FF).
   
   WORD wSOIMarker;
-  int n = fread(&wSOIMarker, 1, 2, f);
+  int n = (int)fread(&wSOIMarker, 1, 2, f);
   if(n!=2)
     return FALSE;
